Add DataTables to /logs.
Datatables helps beautify sorting, searching etc. on tables. The table has been configured to make ajax request to the server and all processing takes place on server-side. This helps handle large amount of data as all data is not sent in one go and browser does not need to handle all data.