Merge filter and filter-having-keys
Both these functions involve filtering in/out data, so it makes sense to
have them in the same tab.

"Filter having keys" is not very clear. Hopefully the new descriptive
sentence ("For the following keys, filter out events that don't have
values") will be a little clearer to new users.