Add remove item link to each value in a filter
Instead of only allowing the user to delete the last value, it's nice
to be able to delete any value in the list.

Style the remove button as a red '-', and add a green '+' to the add
button to make this clear.

Also, move the add button above the checkbox, so it's clear that it's
referring to the list of values, while the checkbox is referring to the
whole filter.