Find all keys from get_explore_data; use it for group pulldown
Here we create two arrays, one of our "recommended" list of group keys,
and one that we get from the get_explore_data API that returns all
the keys we have access to.

Using a checkbox, the user can have the pulldown show either the recommended
list or the entire list.