Explicitly allow log mgr to be iframed in CODAP
Rails 4 sets the X-Frame-Options header to “SAMEORIGIN”, which previously did not matter because we used to overwrite all the default headers in config/application.rb (in order to set custom Access-Control headers).

When I added the rack-cors gem, I got rid of the header override. But this meant that the default X-Frame-Options header was served, and browsers running CODAP would refuse to load the log manager.