Change log creation to work w/previous migration
Parameters now needs to be {} if nil due to NOT NULL constraint

Extras now needs to be assigned to its value in one step. Assigning Log.extras an empty hash, and then modifying the hash (as was done previously) no longer marks extras dirty/savable because the empty hash is the default value on creation.